# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chatbot & Interview Copilot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface User {
  id: string;
  /** @nullable */
  name?: string | null;
  email: string;
  /** @nullable */
  image?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SignupInput {
  name?: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
}

export type BotType = typeof BotType[keyof typeof BotType];


export const BotType = {
  CHATBOT: 'CHATBOT',
  INTERVIEW_COPILOT: 'INTERVIEW_COPILOT',
} as const;

export interface Bot {
  id: string;
  userId: string;
  type: BotType;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  jobDescription?: string | null;
  /** @nullable */
  answerStyle?: string | null;
  temperature: number;
  createdAt: string;
  updatedAt: string;
}

export interface BotInput {
  type: BotType;
  name: string;
  description?: string;
  systemPrompt?: string;
  jobDescription?: string;
  answerStyle?: string;
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
}

export interface BotUpdate {
  name?: string;
  description?: string;
  systemPrompt?: string;
  jobDescription?: string;
  answerStyle?: string;
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
}

export type DocumentVectorStatus = typeof DocumentVectorStatus[keyof typeof DocumentVectorStatus];


export const DocumentVectorStatus = {
  pending: 'pending',
  processing: 'processing',
  processed: 'processed',
  failed: 'failed',
} as const;

export interface Document {
  id: string;
  botId: string;
  userId: string;
  filename: string;
  originalName: string;
  mimeType: string;
  fileSize: number;
  /** @nullable */
  chunkCount?: number | null;
  vectorStatus: DocumentVectorStatus;
  uploadedAt: string;
}

export interface ChatMessageInput {
  content: string;
  /** @nullable */
  conversationId?: string | null;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  USER: 'USER',
  ASSISTANT: 'ASSISTANT',
} as const;

export interface Message {
  id: string;
  conversationId: string;
  role: MessageRole;
  content: string;
  /** @nullable */
  sourceDocs?: string | null;
  createdAt: string;
}

export interface Conversation {
  id: string;
  userId: string;
  botId: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  botType?: string | null;
  /** @nullable */
  botName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationWithMessages {
  id: string;
  userId: string;
  botId: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  botType?: string | null;
  /** @nullable */
  botName?: string | null;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface ConversationInput {
  botId: string;
  title?: string;
}

export interface ConversationUpdate {
  title?: string;
}

export interface ProfileUpdate {
  name?: string;
  image?: string;
}