# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Chatbot & Interview Copilot API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: bots
    description: Bot management
  - name: documents
    description: Document management
  - name: conversations
    description: Conversation management
  - name: messages
    description: Message management
  - name: settings
    description: User settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List all bots for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"

    post:
      operationId: createBot
      tags: [bots]
      summary: Create a new bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

  /bots/{botId}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get a bot by ID
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateBot
      tags: [bots]
      summary: Update a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content

  /bots/{botId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document to a bot (multipart handled by server directly)
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /bots/{botId}/documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content

  /bots/{botId}/chat:
    post:
      operationId: sendChatMessage
      tags: [bots]
      summary: Send a message to a bot (SSE streaming)
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Streamed response
          content:
            text/event-stream:
              schema:
                type: string

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List all conversations for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{conversationId}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a conversation with messages
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateConversation
      tags: [conversations]
      summary: Update a conversation title
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

    delete:
      operationId: deleteConversation
      tags: [conversations]
      summary: Delete a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content

  /settings/profile:
    patch:
      operationId: updateProfile
      tags: [settings]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /settings/delete-chats:
    post:
      operationId: deleteAllChats
      tags: [settings]
      summary: Delete all conversations for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /settings/delete-documents:
    post:
      operationId: deleteAllDocuments
      tags: [settings]
      summary: Delete all documents for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /settings/account:
    delete:
      operationId: deleteAccount
      tags: [settings]
      summary: Delete user account
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    User:
      type: object
      required: [id, email, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: string
        image:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    SignupInput:
      type: object
      required: [email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    BotType:
      type: string
      enum: [CHATBOT, INTERVIEW_COPILOT]

    Bot:
      type: object
      required: [id, userId, type, name, temperature, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          $ref: "#/components/schemas/BotType"
        name:
          type: string
        description:
          type: ["string", "null"]
        systemPrompt:
          type: ["string", "null"]
        jobDescription:
          type: ["string", "null"]
        answerStyle:
          type: ["string", "null"]
        temperature:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    BotInput:
      type: object
      required: [type, name]
      properties:
        type:
          $ref: "#/components/schemas/BotType"
        name:
          type: string
        description:
          type: string
        systemPrompt:
          type: string
        jobDescription:
          type: string
        answerStyle:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2

    BotUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        systemPrompt:
          type: string
        jobDescription:
          type: string
        answerStyle:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2

    Document:
      type: object
      required: [id, botId, userId, filename, originalName, mimeType, fileSize, vectorStatus, uploadedAt]
      properties:
        id:
          type: string
        botId:
          type: string
        userId:
          type: string
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        chunkCount:
          type: ["integer", "null"]
        vectorStatus:
          type: string
          enum: [pending, processing, processed, failed]
        uploadedAt:
          type: string
          format: date-time

    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        conversationId:
          type: ["string", "null"]

    Message:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: string
        conversationId:
          type: string
        role:
          type: string
          enum: [USER, ASSISTANT]
        content:
          type: string
        sourceDocs:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    Conversation:
      type: object
      required: [id, userId, botId, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        botId:
          type: string
        title:
          type: ["string", "null"]
        botType:
          type: ["string", "null"]
        botName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ConversationWithMessages:
      type: object
      required: [id, userId, botId, createdAt, updatedAt, messages]
      properties:
        id:
          type: string
        userId:
          type: string
        botId:
          type: string
        title:
          type: ["string", "null"]
        botType:
          type: ["string", "null"]
        botName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"

    ConversationInput:
      type: object
      required: [botId]
      properties:
        botId:
          type: string
        title:
          type: string

    ConversationUpdate:
      type: object
      properties:
        title:
          type: string

    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        image:
          type: string