# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chatbot & Interview Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  content: string;
  /** @nullable */
  conversationId?: string | null;
}